Add files via upload
A program that takes in the name of items and makes use of vectors. It has a functionality to add and remove an item after finalization. It cannot take a string with spaces after you choose to add an item to a final list (this was the only bug I could find).